# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/openauth/permissions": {
            "get": {
                "description": "Get a list of all permissions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Get all permissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "AuthToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit the number of results (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetPermissionsResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new permission with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Create a new permission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "AuthToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Permission details",
                        "name": "permission",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.PermissionDetails"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a permission by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Delete a permission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "AuthToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Permission details",
                        "name": "permission",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeletePermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permission deleted successfully"
                    }
                }
            }
        },
        "/openauth/permissions/groups": {
            "get": {
                "description": "Get permissions of a group by group ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Get permissions of a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "AuthToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Group ID",
                        "name": "groupId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PermissionDetailsShort"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add permissions to a group by group ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Add permissions to a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "AuthToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Permissions details",
                        "name": "permissions",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddRemovePermissionsToGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permissions added to group successfully"
                    }
                }
            },
            "delete": {
                "description": "Remove permissions from a group by group ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Remove permissions from a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "AuthToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Permissions details",
                        "name": "permissions",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddRemovePermissionsToGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permissions removed from group successfully"
                    }
                }
            }
        },
        "/openauth/permissions/users": {
            "get": {
                "description": "Get permissions of a user by user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Get permissions of a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "AuthToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PermissionDetailsShort"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add permissions to a user by user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Add permissions to a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "AuthToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Permissions details",
                        "name": "permissions",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddRemovePermissionsToUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permissions added to user successfully"
                    }
                }
            },
            "delete": {
                "description": "Remove permissions from a user by user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Remove permissions from a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "AuthToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Permissions details",
                        "name": "permissions",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddRemovePermissionsToUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permissions removed from user successfully"
                    }
                }
            }
        },
        "/openauth/permissions/{id}": {
            "get": {
                "description": "Get details of a permission by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Get details of a permission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "AuthToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Permission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PermissionDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AddRemovePermissionsToGroupRequest": {
            "type": "object",
            "properties": {
                "groupId": {
                    "type": "integer"
                },
                "permIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "dto.AddRemovePermissionsToUserRequest": {
            "type": "object",
            "properties": {
                "permIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CreatePermissionRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.DeletePermissionRequest": {
            "type": "object",
            "properties": {
                "permId": {
                    "type": "integer"
                }
            }
        },
        "dto.GetPermissionsResponse": {
            "type": "object",
            "properties": {
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PermissionDetails"
                    }
                }
            }
        },
...
        "dto.PermissionDetails": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "createdByUser": {
                    "type": "integer"
                },
                "createdOn": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedOn": {
                    "type": "integer"
                }
            }
        },
        "dto.PermissionDetailsShort": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}